chore(backend): rename errors file to route-errors
